.ipynb_checkpoints/requirements-checkpoint.txt
absl-py==0.12.0
adal==1.2.7
adversarial-robustness-toolbox==1.6.1
aiohttp==3.7.4
ansiwrap==0.8.4
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.0.3
asn1crypto==1.4.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
binaryornot==0.4.4
black==20.8b1
bleach==3.3.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.1
caip-notebooks-serverextension==1.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cma==3.0.3
colorama==0.4.4
conda==4.9.2
conda-package-handling==1.7.3
confuse==1.4.0
cookiecutter==1.7.2
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.6
defusedxml==0.7.1
docker==5.0.0
docker-pycreds==0.4.0
entrypoints==0.3
ffmpeg-python==0.2.0
fsspec==0.9.0
future==0.18.2
gcsfs==0.8.0
gitdb==4.0.7
GitPython==3.1.14
google-api-core==1.26.2
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
google-cloud-bigquery==2.13.1
google-cloud-bigquery-storage==2.4.0
google-cloud-bigtable==2.0.0
google-cloud-core==1.6.0
google-cloud-dataproc==2.3.1
google-cloud-datastore==2.1.0
google-cloud-firestore==2.1.0
google-cloud-kms==2.2.0
google-cloud-language==2.0.0
google-cloud-logging==2.3.1
google-cloud-monitoring==2.2.1
google-cloud-pubsub==1.7.0
google-cloud-scheduler==2.2.0
google-cloud-spanner==3.3.0
google-cloud-speech==2.2.1
google-cloud-storage==1.37.1
google-cloud-tasks==2.2.0
google-cloud-translate==3.1.0
google-cloud-videointelligence==2.1.0
google-cloud-vision==2.3.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpc-google-iam-v1==0.12.3
grpcio==1.37.0
grpcio-gcp==0.2.2
htmlmin==0.1.12
httplib2==0.19.1
idna==2.10
ImageHash==4.2.0
importlib-metadata==3.10.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-http-over-ws==0.0.8
jupyterlab==1.2.16
jupyterlab-git==0.11.0
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kubernetes==12.0.1
libcst==0.3.18
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
missingno==0.4.2
mistune==0.8.4
multidict==5.1.0
mypy==0.812
mypy-extensions==0.4.3
nb-conda==2.2.1
nb-conda-kernels==2.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbdime==2.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.3.0
notebook-executor==0.2
numba==0.53.1
numpy==1.19.5
oauthlib==3.0.1
olefile==0.46
packaging==20.9
pandas==1.2.4
pandas-profiling==2.11.0
pandocfilters==1.4.2
papermill==2.3.3
parso==0.8.2
pathspec==0.8.1
patsy==0.5.1
pexpect==4.8.0
phik==0.11.2
pickleshare==0.7.5
Pillow==8.1.2
pip==21.0.1
poyo==0.5.0
prettytable==2.1.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.15.8
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.3
pycparser==2.20
pydub==0.25.1
Pygments==2.8.1
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytorch-ignite==0.4.4
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rsa==4.7.2
ruamel-yaml-conda==0.15.80
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.5.0
setuptools==49.6.0.post20210108
simplejson==3.17.2
six==1.15.0
smmap==3.0.5
SQLAlchemy==1.4.7
sqlparse==0.4.1
statsmodels==0.12.2
tangled-up-in-unicode==0.0.7
tenacity==7.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
textwrap3==0.9.2
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.0
torchvision==0.9.0+cu111
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
ujson==4.0.2
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.4
visions==0.6.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wheel==0.36.2
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
